refactor(desktop): migrate FearAndGreed to standalone MVVM feature
Extract FearAndGreed from MarketBanner into its own feature folder,
following the MVVM architecture pattern with a dedicated ViewModel hook.
