+++ .changeset/gorgeous-beers-bake.md
@@ -1,0 +1,5 @@
+---
+"ledger-live-desktop": minor
+---
+
+Migrate FearAndGreed as a feature
