+++ apps/ledger-live-desktop/src/mvvm/features/FearAndGreed/components/LoadingTile.tsx
