+++ apps/ledger-live-desktop/src/mvvm/features/FearAndGreed/hooks/useFearAndGreedViewModel.ts
