+++ apps/ledger-live-desktop/src/mvvm/features/MarketBanner/components/TrendingAssetsList.tsx
@@ -4,10 +4,10 @@ import { MarketItemPerformer } from "@ledgerhq/live-common/market/utils/types";
 import { PerformanceIndicator } from "./PerformanceIndicator";
 import { useNavigate } from "react-router";
 import { ViewAllTile } from "./ViewAllTile";
-import FearAndGreed from "./FearAndGreed";
 import { AssetIcon } from "./AssetIcon";
 import { track } from "~/renderer/analytics/segment";
 import { TRACKING_PAGE_NAME } from "../utils/constants";
+import FearAndGreed from "LLD/features/FearAndGreed";
 
 type TrendingAssetsListProps = {
   readonly items: MarketItemPerformer[];
